# docs/swagger.yaml
definitions:
  dto.CaptchaVerify:
    properties:
      code:
        type: string
      id:
        type: string
    required:
    - code
    - id
    type: object
...
  dto.Pagination:
    properties:
      current:
        type: integer
      pageSize:
        type: integer
      total:
        type: integer
    type: object
  echox.Response:
    properties:
      data: {}
      message: {}
    type: object
  models.Menu:
    properties:
      actions:
        items:
          $ref: '#/definitions/models.MenuAction'
        type: array
      component:
        type: string
      created_at:
        $ref: '#/definitions/sql.NullTime'
      created_by:
        type: string
      deleted:
        type: boolean
      hidden:
        maximum: 1
        minimum: -1
        type: integer
      icon:
        type: string
      id:
        type: string
      name:
        type: string
      parent_id:
        type: string
      parent_path:
        type: string
      remark:
        type: string
      router:
        type: string
      sequence:
        type: integer
      status:
        maximum: 1
        minimum: -1
        type: integer
      updated_at:
        $ref: '#/definitions/sql.NullTime'
    required:
    - hidden
    - icon
    - name
    - remark
    - sequence
    - status
    type: object
  models.MenuAction:
    properties:
      code:
        type: string
      created_at:
        $ref: '#/definitions/sql.NullTime'
      deleted:
        type: boolean
      id:
        type: string
      menu_id:
        type: string
      name:
        type: string
      resources:
        items:
          $ref: '#/definitions/models.MenuActionResource'
        type: array
      updated_at:
        $ref: '#/definitions/sql.NullTime'
    required:
    - code
    - name
    type: object
  models.MenuActionResource:
    properties:
      created_at:
        $ref: '#/definitions/sql.NullTime'
      deleted:
        type: boolean
      method:
        type: string
      path:
        type: string
      updated_at:
        $ref: '#/definitions/sql.NullTime'
    required:
    - method
    - path
    type: object
  models.MenuQueryResult:
    properties:
      list:
        items:
          $ref: '#/definitions/models.Menu'
        type: array
      pagination:
        $ref: '#/definitions/dto.Pagination'
    type: object
...
  sql.NullTime:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
info:
  contact: {}
paths:
  /api/v1/menus:
    get:
      parameters:
      - in: query
        name: current
        type: integer
      - enum:
        - ASC
        - DESC
        in: query
        name: direction
        type: string
        x-enum-varnames:
        - OrderByASC
        - OrderByDESC
      - in: query
        maximum: 1
        minimum: -1
        name: hidden
        type: integer
      - collectionFormat: csv
        in: query
        items:
          type: string
        name: ids
        type: array
      - in: query
        name: includeActions
        type: boolean
      - in: query
        name: key
        type: string
      - in: query
        name: name
        type: string
      - in: query
        maximum: 128
        name: pageSize
        type: integer
      - in: query
        name: parentID
        type: string
      - in: query
        name: prefixParentPath
        type: string
      - in: query
        name: queryValue
        type: string
      - in: query
        maximum: 1
        minimum: -1
        name: status
        type: integer
      - in: query
        name: tree
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            allOf:
            - $ref: '#/definitions/echox.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.MenuQueryResult'
              type: object
        "400":
          description: bad request
          schema:
            type: string
        "500":
          description: internal error
          schema:
            type: string
      summary: Query
      tags:
      - Menu
    post:
      parameters:
      - description: Menu
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.Menu'
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            $ref: '#/definitions/echox.Response'
        "400":
          description: bad request
          schema:
            $ref: '#/definitions/echox.Response'
        "500":
          description: internal error
          schema:
            $ref: '#/definitions/echox.Response'
      summary: Menu Create
      tags:
      - Menu
  /api/v1/menus/{id}:
    delete:
      parameters:
      - description: menu id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            $ref: '#/definitions/echox.Response'
        "400":
          description: bad request
          schema:
            $ref: '#/definitions/echox.Response'
        "500":
          description: internal error
          schema:
            $ref: '#/definitions/echox.Response'
      summary: Menu Delete By ID
      tags:
      - Menu
    get:
      parameters:
      - description: menu id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            allOf:
            - $ref: '#/definitions/echox.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.Menu'
              type: object
        "400":
          description: bad request
          schema:
            $ref: '#/definitions/echox.Response'
        "500":
          description: internal error
          schema:
            $ref: '#/definitions/echox.Response'
      summary: Menu Get By ID
      tags:
      - Menu
    put:
      parameters:
      - description: menu id
        in: path
        name: id
        required: true
        type: integer
      - description: Menu
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.Menu'
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            $ref: '#/definitions/echox.Response'
        "400":
          description: bad request
          schema:
            $ref: '#/definitions/echox.Response'
        "500":
          description: internal error
          schema:
            $ref: '#/definitions/echox.Response'
      summary: Menu Update By ID
      tags:
      - Menu
  /api/v1/menus/{id}/actions:
    get:
      parameters:
      - description: menu id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            allOf:
            - $ref: '#/definitions/echox.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.MenuAction'
                  type: array
              type: object
        "400":
          description: bad request
          schema:
            $ref: '#/definitions/echox.Response'
        "500":
          description: internal error
          schema:
            $ref: '#/definitions/echox.Response'
      summary: MenuActions Get By ID
      tags:
      - Menu
    put:
      parameters:
      - description: menu id
        in: path
        name: id
        required: true
        type: integer
      - description: MenuActions
        in: body
        name: data
        required: true
        schema:
          items:
            $ref: '#/definitions/models.MenuAction'
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            $ref: '#/definitions/echox.Response'
        "400":
          description: bad request
          schema:
            $ref: '#/definitions/echox.Response'
        "500":
          description: internal error
          schema:
            $ref: '#/definitions/echox.Response'
      summary: MenuActions Update By ID
      tags:
      - Menu
  /api/v1/menus/{id}/disable:
    patch:
      parameters:
      - description: menu id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            $ref: '#/definitions/echox.Response'
        "400":
          description: bad request
          schema:
            $ref: '#/definitions/echox.Response'
        "500":
          description: internal error
          schema:
            $ref: '#/definitions/echox.Response'
      summary: Menu Disable By ID
      tags:
      - Menu
  /api/v1/menus/{id}/enable:
    patch:
      parameters:
      - description: menu id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            $ref: '#/definitions/echox.Response'
        "400":
          description: bad request
          schema:
            $ref: '#/definitions/echox.Response'
        "500":
          description: internal error
          schema:
            $ref: '#/definitions/echox.Response'
      summary: Menu Enable By ID
      tags:
      - Menu
  /api/v1/publics/captcha:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            type: string
        "400":
          description: bad request
          schema:
            type: string
        "500":
          description: internal error
          schema:
            type: string
      summary: GetCaptcha
      tags:
      - Public
  /api/v1/publics/captcha/verify:
    post:
      parameters:
      - description: CaptchaVerify
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.CaptchaVerify'
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            type: string
        "400":
          description: bad request
          schema:
            type: string
      summary: VerifyCaptcha
      tags:
      - Public
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"